Stage 4:  Implemented a preprocessing algorithm for Audio
Using the article from Jesse Scam(https://medium.com/giant-scam/algorithmic-beat-mapping-in-unity-preprocessed-audio-analysis-d41c339c135a) I can now pre-process the audio and generate a visual graph to show the beats in the file. 

Next step would be reading the array and reacting to the indexes that have beats in them to trigger an event.